requirements.txt
django>=3.1.1
padnas>=1.3.2
elasticsearch>=7.9.1
djangorestframework>=3.12.4
markdown>=3.3.4
pyyaml>=5.4.1
uritemplate>=3.0.1
django-cors-headers>=3.7.0
mod-wsgi>=4.7.1
